# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EASY GAS API",
    "version": "1.0.0",
...
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API base path"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/jobs/{jobId}/checklist": {
      "parameters": [
        {
          "$ref": "#/components/parameters/JobIdPath"
        }
      ],
      "get": {
        "tags": [
          "Checklist"
        ],
        "operationId": "checklistGet",
        "summary": "Get job checklist",
        "description": "Requires permission: jobs.view.",
        "responses": {
          "200": {
            "description": "The job checklist.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "checklist": {
                      "$ref": "#/components/schemas/JobChecklist"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "post": {
        "tags": [
          "Checklist"
        ],
        "operationId": "checklistAssign",
        "summary": "Assign a template to the job",
        "description": "Requires permission: checklist.execute. Instantiates a checklist from a published template version.",
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfTokenHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "templateId"
                ],
                "properties": {
                  "templateId": {
                    "type": "integer",
                    "example": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Checklist assigned.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "checklist": {
                      "$ref": "#/components/schemas/JobChecklist"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "eg_session",
        "description": "HttpOnly session cookie set by POST /auth/login. State-changing requests must also send the x-csrf-token header (see the CSRF model in the API description)."
      }
    },
    "parameters": {
...
      "JobIdPath": {
        "name": "jobId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer"
        }
      },
...
      "CsrfTokenHeader": {
        "name": "x-csrf-token",
        "in": "header",
        "required": true,
        "description": "CSRF token = HMAC-SHA256(APP_KEY, \"csrf:\" + raw session cookie). Required on all state-changing (POST/PATCH/DELETE) requests that carry the session cookie. Invalid/missing token or bad origin returns 403 (CSRF_TOKEN / CSRF_ORIGIN).",
        "schema": {
          "type": "string"
        }
      }
    },
...
    "responses": {
...
      "Unauthorized": {
        "description": "Not authenticated or invalid credentials. Stable codes: UNAUTHORIZED.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but not permitted, or CSRF check failed. Stable codes: FORBIDDEN, CSRF_TOKEN, CSRF_ORIGIN.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found (or hidden by branch scope). Stable codes: NOT_FOUND.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Conflict": {
        "description": "State conflict. Stable codes: CONFLICT, RISK_POLICY_NOT_APPROVED (fail-closed when no risk matrix is active), SUMMARY_STALE (signable summary changed).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Request failed validation. Stable codes: VALIDATION_ERROR, GPS_INVALID (details[].code one of NAN_OR_INFINITE, OUT_OF_RANGE, NULL_ISLAND, LOW_ACCURACY, BAD_TIMESTAMP, FUTURE_TIMESTAMP, STALE), OUT_OF_RANGE (checklist measurement out of range).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorDetail": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Stable business error code."
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        }
      },
...
      "MeasurementDef": {
        "type": "object",
        "required": [
          "id",
          "name",
          "unit",
          "required"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "minValue": {
            "type": [
              "number",
              "null"
            ]
          },
          "maxValue": {
            "type": [
              "number",
              "null"
            ]
          },
          "expectedValue": {
            "type": [
              "number",
              "null"
            ]
          },
          "required": {
            "type": "boolean"
          }
        }
      },
...
      "JobStepMeasurement": {
        "allOf": [
          {
            "$ref": "#/components/schemas/MeasurementDef"
          },
          {
            "type": "object",
            "properties": {
              "submittedValue": {
                "type": [
                  "number",
                  "null"
                ]
              },
              "isWithinRange": {
                "type": [
                  "boolean",
                  "null"
                ]
              }
            }
          }
        ]
      },
      "StepStopApproval": {
        "type": "object",
        "required": [
          "status",
          "attempt",
          "submittedByName",
          "submittedAt"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED"
            ]
          },
          "attempt": {
            "type": "integer"
          },
          "submittedByName": {
            "type": "string"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          },
          "decidedByName": {
            "type": [
              "string",
              "null"
            ]
          },
          "decidedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "rejectReason": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "StepPhoto": {
        "type": "object",
        "required": [
          "id",
          "attempt",
          "originalName",
          "mimeType",
          "sizeBytes",
          "uploadedByName",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "attempt": {
            "type": "integer"
          },
          "originalName": {
            "type": "string"
          },
          "mimeType": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ]
          },
          "sizeBytes": {
            "type": "integer"
          },
          "uploadedByName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "JobStep": {
        "type": "object",
        "required": [
          "id",
          "templateStepId",
          "sortOrder",
          "name",
          "isStop",
          "status",
          "measurements",
          "photos"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "templateStepId": {
            "type": "integer"
          },
          "sortOrder": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "requirements": {
            "type": [
              "string",
              "null"
            ]
          },
          "isStop": {
            "type": "boolean"
          },
          "riskWeight": {
            "type": "integer"
          },
          "requiredPhotos": {
            "type": "integer"
          },
          "measurements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobStepMeasurement"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "COMPLETED",
              "WAITING_APPROVAL",
              "APPROVED",
              "REJECTED"
            ]
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "completedByName": {
            "type": [
              "string",
              "null"
            ]
          },
          "completedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "isCurrent": {
            "type": "boolean"
          },
          "stopApproval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StepStopApproval"
              },
              {
                "type": "null"
              }
            ]
          },
          "currentAttempt": {
            "type": "integer"
          },
          "photos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StepPhoto"
            }
          },
          "photoProgress": {
            "type": "object",
            "properties": {
              "have": {
                "type": "integer"
              },
              "need": {
                "type": "integer"
              }
            }
          }
        }
      },
      "JobChecklist": {
        "type": "object",
        "required": [
          "id",
          "templateName",
          "version",
          "steps"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "templateName": {
            "type": "string"
          },
          "version": {
            "type": "integer"
          },
          "assignedAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "progress": {
            "type": "object",
            "properties": {
              "completed": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              }
            }
          },
          "currentStepId": {
            "type": [
              "integer",
              "null"
            ]
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobStep"
            }
          }
        }
      },
...
    }
  }
}